code: переход на long

+++ main.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-	unsigned int a, b;
+	long a, b;
 	scanf("%d %d", &a, &b);
 	printf("a + b = %d", a + b);
 	printf("a + b = %d", a - b);
